Sprint 1 — LLM 层重构:
- types/llm 基础类型统一和响应工具
- Provider 通用基类抽取和工厂模式
- 中间件泛型化 (circuitbreaker/idempotency/retry)
- multimodal/embedding/streaming/tools/router 清理

Sprint 2 — RAG + Workflow 层:
- DocumentLoader 统一接口 + 5 种格式实现
- Config→RAG 桥接工厂 + float32/64 转换层
- Workflow 步骤集成 + DSL 表达式引擎

Sprint 3 — Agent 核心重构:
- base/builder/completion/react 核心模块优化
- MCP WebSocket 传输 + Streaming 适配器
- 插件系统 + 声明式 Agent 定义

Sprint 5 — P0 新特性:
- Agent-as-Tool 适配器（多智能体组合）
- RunConfig 运行时配置覆盖
- Guardrails Tripwire 熔断 + Parallel 并行
- Context Window 自动管理（3 种策略）

代码规范: §18-§23 可执行契约沉淀

+++ rag/graph_rag.go
@@ -16,7 +16,7 @@ type Node struct {
 	Type       string         `json:"type"`
 	Label      string         `json:"label"`
 	Properties map[string]any `json:"properties,omitempty"`
-	Embedding  []float32      `json:"embedding,omitempty"`
+	Embedding  []float64      `json:"embedding,omitempty"`
 	CreatedAt  time.Time      `json:"created_at"`
 }
 
@@ -155,8 +155,8 @@ type GraphRAG struct {
 
 // GraphRAG中矢量操作的 GraphVectorStore 接口.
 type GraphVectorStore interface {
-	Store(ctx context.Context, id string, embedding []float32, metadata map[string]any) error
-	Search(ctx context.Context, embedding []float32, limit int) ([]GraphVectorResult, error)
+	Store(ctx context.Context, id string, embedding []float64, metadata map[string]any) error
+	Search(ctx context.Context, embedding []float64, limit int) ([]GraphVectorResult, error)
 }
 
 // GraphVectorResult 代表着向量搜索结果.
@@ -168,7 +168,7 @@ type GraphVectorResult struct {
 
 // GraphEmbedder 生成嵌入式.
 type GraphEmbedder interface {
-	Embed(ctx context.Context, text string) ([]float32, error)
+	Embed(ctx context.Context, text string) ([]float64, error)
 }
 
 // GraphRAGConfig 配置了 GraphRAG.
